feat: activityにあった実装をFragmentに移しました。
