+++ app/src/main/res/layout/activity_message.xml
@@ -3,9 +3,7 @@
         xmlns:tools="http://schemas.android.com/tools">
     <data>
         <import type="jp.panta.misskeyandroidclient.ui.SafeUnbox"/>
-        <variable
-                name="actionViewModel"
-                type="jp.panta.misskeyandroidclient.ui.messaging.viewmodel.MessageActionViewModel" />
+
     </data>
     <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
 
@@ -32,80 +30,10 @@
                 app:layout_constraintHorizontal_bias="0.0"
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toBottomOf="@+id/messageToolbar"
-                app:layout_constraintBottom_toTopOf="@id/userAction"
+                app:layout_constraintBottom_toBottomOf="parent"
                 app:layout_constraintVertical_bias="0.0">
 
         </FrameLayout>
-        <LinearLayout
-                android:id="@+id/userAction"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/content_main"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
-                android:orientation="vertical"
-                android:padding="8dp"
-                >
 
-
-            <View
-                    android:layout_width="match_parent"
-                    android:layout_height="0.1dp"
-                    android:background="?attr/colorSeparator"
-                    />
-            <MultiAutoCompleteTextView
-                    android:id="@+id/inputMessage"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:ems="10"
-                    android:inputType="textMultiLine"
-                    android:hint="@string/input_message"
-                    android:text="@={actionViewModel.text}"
-                    />
-
-            <LinearLayout
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:orientation="horizontal"
-                    >
-
-
-                <ImageButton
-                        android:id="@+id/openDrive"
-                        android:layout_width="48dp"
-                        android:layout_height="48dp"
-                        android:padding="8dp"
-                        app:srcCompat="@drawable/ic_cloud_black_24dp"
-                        style="@style/Widget.AppCompat.Button.Borderless"
-                        tools:ignore="ContentDescription"
-                        app:tint="?attr/normalIconTint" />
-                <TextView
-                        android:id="@+id/selecting_file_name"
-                        android:layout_width="0dp"
-                        android:layout_height="wrap_content"
-                        tools:text="panta.jpg"
-                        android:layout_gravity="center"
-                        android:text="@{SafeUnbox.unbox(actionViewModel.file.name)}"
-                        android:layout_weight="1"
-                        >
-
-                </TextView>
-                <ImageButton
-                        android:layout_gravity="end"
-                        android:id="@+id/sendButton"
-                        android:layout_width="48dp"
-                        android:layout_height="48dp"
-                        android:padding="8dp"
-
-                        app:srcCompat="@drawable/ic_menu_send"
-                        style="@style/Widget.AppCompat.Button.Borderless"
-                        tools:ignore="ContentDescription"
-                        android:onClick="@{()-> actionViewModel.send()}"
-                        app:tint="?attr/normalIconTint" />
-
-
-            </LinearLayout>
-        </LinearLayout>
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
